requirements.txt
fastapi
uvicorn[standard]
python-dotenv
retell-sdk
twilio
sendgrid